+++ GRDB/Core/StatementColumnConvertible.swift
@@ -223,7 +223,7 @@ where Value: DatabaseValueConvertible & StatementColumnConvertible
         }
         
         // Assume cursor is created for immediate iteration: reset and set arguments
-        try statement.reset(withArguments: arguments)
+        try statement.prepareExecution(withArguments: arguments)
     }
     
     deinit {